fastapi_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
loguru==0.7.2
numpy==1.26.4
Pillow==10.4.0
SQLAlchemy==2.0.32
pandas==2.2.2
tensorflow-cpu==2.16.1
scikit-learn==1.5.1
mlflow==2.15.1
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.3.2
